skip users after timeout
This enables a 2 minute timeout for users who have received the
notification but have not acted on it. The basic idea is we launch a
task with a 2 minute countdown to send the buttons to someone else. In
the meantime we track who the buttons were last sent to, and when the
task runs it checks to see if the first user has in the meantime
pressed "Reject". If everyone is skipped then further action is taken
and anyone can press "Accept". In the future this could be made more
efficient.

Also, this substitutes the expiration message to the requestor with a
message when all available users have been notified but no one has yet
accepted.

Last, we make the redis hostname configurable, separated the celery
app from the tasks (so that the tasks can be loaded lazily by celery),
named the tasks.